float

+++ Tests/Hello.cpp
@@ -10,8 +10,8 @@ using namespace std;
 // Define function that runs on the GPU.
 
 // out 1132 vec 5000 main 633
-std::vector<double> vec = {1, 2, 3, 4};
-std::vector<double> main_filter = {5, 6, 7};
+std::vector<float> vec = {1, 2, 3, 4};
+std::vector<float> main_filter = {5, 6, 7};
 static const int out_siz = vec.size() + main_filter.size() - 1;
 SharedArray<float> out(out_siz);
 
@@ -19,7 +19,7 @@ static int const main_siz = main_filter.size();
 static int const vec_siz = vec.size();
 
 
- inline void multi_vec_elem(double elem, int it) {
+ inline void multi_vec_elem(float elem, int it) {
      for(int i = 0; i < main_siz ; i++){
         out[it] += (float) main_filter[i] * elem;
         it++;
